+++ packages/design-system/vitest.config.js
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-import baseConfig from '@hypergiant/vitest-config/dom';
+import baseConfig from '@accelint/vitest-config/dom';
 import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
 import { defineConfig, mergeConfig } from 'vitest/config';