+++ packages/design-system/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @hypergiant/design-system
+# @accelint/design-system
 
 ## 0.1.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [eba7ce9]
-  - @hypergiant/converters@0.1.0
+  - @accelint/converters@0.1.0