+++ tooling/eslint-config/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @hypergiant/eslint-config
+# @accelint/eslint-config
 
 ## 0.1.0